feature: restructure project to avoid clutter - containers: stateful Components, companion css files in here also - components: stateless Components - images: replace icons in generated structure - rewire routes and references accordingly
